MOCK_BPF: Remove bpf mocking library
Removing the bpf mocking system because we feel that it is not
useful in properly testing bpf programs. There is currently only one
real user of it (time_in_state_test.cpp), which has significant issues
in that the tests are written in such a way that each one relies on the
state of the previous test. Also, the per CPU mocking capability is
limited in its capabilities.

Removal simplifies the transition to libbpf, and reduces legacy tech
debt.

Bug: 359646531
Test: TH
Change-Id: I402a6dec8e8b750161bf36fea2340cf54a1b7433
Signed-off-by: Neill Kapron <nkapron@google.com>